fix: move IdentityInitializationException to security-validation module
