+++ qa/acceptance-tests/src/test/java/io/camunda/it/identity/InitializeInvalidAuthorizationIT.java
@@ -10,7 +10,7 @@ package io.camunda.it.identity;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import io.camunda.security.configuration.ConfiguredAuthorization;
-import io.camunda.zeebe.engine.processing.identity.initialize.IdentityInitializationException;
+import io.camunda.security.validation.IdentityInitializationException;
 import io.camunda.zeebe.protocol.record.value.AuthorizationOwnerType;
 import io.camunda.zeebe.protocol.record.value.PermissionType;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
